refactor: Move AmaiSosu.Installation.IO namespace to AmaiSosu.Common.IO
