+++ src/AmaiSosu/Main.Compile.cs
@@ -22,8 +22,8 @@ using System;
 using System.Diagnostics;
 using System.IO;
 using System.Reflection;
+using AmaiSosu.Common.IO;
 using AmaiSosu.Installation;
-using AmaiSosu.Installation.IO;
 using static System.Environment;
 using static AmaiSosu.Common.Paths;
 using static AmaiSosu.Resources.FileNames;
